requirements.txt
google-api-python-client
google-genai
pandas
python-dotenv